Version 2

+++ script.js
@@ -10,39 +10,44 @@ const handleSubmit = (event) => {
     event.preventDefault();
 
     const inputValue = document.getElementById("todo-input").value;
+
+    if(inputValue === ""){
+        window.alert("Bitte einen Wert eingeben!");
+    } else {
+
     const newListItem = document.createElement("li");
 
     newListItem.innerHTML = `<div class='row mt-5'>
                             <div class='col-9'><input type='text' class='w-100' name='todo' value='${inputValue}'></input></div>
-                            <div class='col-3 text-end'><button type='button' class='btn btn-success'>Eintrag ändern</button>
-                            <button type='button' class='btn btn-danger del'><i class="fa fa-trash" aria-hidden="true"></i> | Eintrag löschen</button></div></div>`;
-    //dem button auch ein name attribute geben (name="del")
+                            <div class='col-3 text-end'><button name='update' type='button' class='btn btn-success'><i class="far fa-edit"></i> | Eintrag ändern</button>
+                            <button type='button' name='delete' class='btn btn-danger del'><i class="far fa-trash-alt"></i> | Eintrag löschen</button></div></div>`;
     list.append(newListItem);
     document.getElementById("todo-input").value = "";
+    }
 };
 
 
-
-//erstelle handleClick (EventHandler)
-//const handleClick = (event) => {
-    //if statement evaluieren ob das e.target, name === "del" || name === "edit"
-    //darauf basierend das event in die jeweilige Funktion (zB zu handleRemove) weiterleiten
- /*   if(event.target.name === "del") {
+const handleClick = (event) => {
+    if(event.target.name === "delete") {
         handleRemove(event)
-    } else 
+    } else if(event.target.name === "update") {
+        let updateToDo = prompt("Neuer Wert für dieses ToDo!");
+        event.target.closest('li').children[0].children[0].children[0].value = updateToDo;
 
+        //document.getElementsByName('todo')[1].value = updateToDo;
+        //console.log(document.getElementsByName('todo')[1]);
+    }    
 }
-*/
+
 
 // Als Arrowfunction 
 const handleRemove = (event) => {
     event.target.closest('li').remove();
 }
 
-
 //EventListener
 form.addEventListener("submit", handleSubmit);
-//list.addEventListener("click", handleClick);
-list.addEventListener("click", handleRemove);  
+list.addEventListener("click", handleClick);
+//list.addEventListener("click", handleRemove);
